sel4vmmcore: Removed double cast

+++ libsel4vmmcore/src/util/io.c
@@ -20,7 +20,7 @@
 
 static int io_port_compare_by_range(const void *pkey, const void *pelem)
 {
-    unsigned int key = (unsigned int)(uintptr_t)pkey;
+    unsigned int key = (unsigned int)pkey;
     const ioport_entry_t *entry = (const ioport_entry_t*)pelem;
     const ioport_range_t *elem = &entry->range;
     if (key < elem->start) {
